Move inline message functions to a class
