+++ src/extension.ts
@@ -2,28 +2,17 @@ import * as vscode from 'vscode';
 import { registerCommands } from './commands';
 import Extension from './components/Extension';
 import StatusBarItem from './components/StatusBarItem';
+import InlineMessage from './components/InlineMessage';
 import { runGitBlameCommand } from './git';
 import {
-  getHoverModalMarkdown,
   getJiraIssueKey,
   getJiraIssueLink,
   JiraViewProvider,
   getJiraIssueContent
 } from './jira';
-import { getInlineMessage } from './utils';
-import { GitBlameCommandInfo } from './types';
 
 let globalContext: vscode.ExtensionContext;
 let webviewProvider: JiraViewProvider;
-
-let inlineMessageEditor: vscode.TextEditor;
-const inlineMessageDecorationType =
-  vscode.window.createTextEditorDecorationType({
-    after: {
-      textDecoration: 'none; opacity: 0.25;',
-      margin: '0 0 0 6em'
-    }
-  });
 
 export function activate(context: vscode.ExtensionContext): void {
   let extension = new Extension(context);
@@ -31,77 +20,6 @@ export function activate(context: vscode.ExtensionContext): void {
   registerCommands(globalContext);
   initWebview();
   bindEventListeners();
-}
-
-// Inline message functions
-async function renderInlineMessage(
-  gitBlameCommandInfo: GitBlameCommandInfo,
-  jiraIssueKey: string
-): Promise<void> {
-  const { gitBlameInfo, editor, lineNumber } = gitBlameCommandInfo;
-  const message = getInlineMessage(gitBlameInfo);
-  // Ensure the line has not changed since running the git command
-  let activeEditor = vscode.window.activeTextEditor;
-  if (
-    !message ||
-    !activeEditor ||
-    activeEditor !== editor ||
-    activeEditor.selection.active.line !== lineNumber
-  ) {
-    hideInlineMessage();
-    return;
-  }
-  inlineMessageEditor = activeEditor;
-  let activeLine = activeEditor.document.lineAt(lineNumber);
-  // Render using the latest information since the length of the line could have changed
-  let range = new vscode.Range(
-    activeLine.lineNumber,
-    activeLine.text.length,
-    activeLine.lineNumber,
-    activeLine.text.length + message.length
-  );
-  const renderOptions = { after: { contentText: message } };
-  let hoverMessage: string | vscode.MarkdownString | undefined;
-  if (jiraIssueKey) {
-    hoverMessage = 'Loading Jira information...';
-  }
-  const decorations = [{ range, renderOptions, hoverMessage }];
-  inlineMessageEditor.setDecorations(inlineMessageDecorationType, decorations);
-  if (jiraIssueKey) {
-    const jiraIssueContent = await getJiraIssueContent(jiraIssueKey);
-    if (jiraIssueContent) {
-      hoverMessage = getHoverModalMarkdown(
-        jiraIssueKey,
-        jiraIssueContent.fields
-      );
-    }
-    // Ensure the line has not changed since running the git command
-    activeEditor = vscode.window.activeTextEditor;
-    if (
-      !activeEditor ||
-      activeEditor !== editor ||
-      activeEditor.selection.active.line !== lineNumber
-    ) {
-      return;
-    }
-    activeLine = activeEditor.document.lineAt(lineNumber);
-    range = new vscode.Range(
-      activeLine.lineNumber,
-      activeLine.text.length,
-      activeLine.lineNumber,
-      activeLine.text.length + message.length
-    );
-    const newDecorations = [{ range, renderOptions, hoverMessage }];
-    inlineMessageEditor.setDecorations(
-      inlineMessageDecorationType,
-      newDecorations
-    );
-  }
-}
-function hideInlineMessage(): void {
-  if (inlineMessageEditor) {
-    inlineMessageEditor.setDecorations(inlineMessageDecorationType, []);
-  }
 }
 
 // Webview functions
@@ -134,24 +52,25 @@ async function renderWebview(jiraIssueKey: string): Promise<void> {
 
 function onChange(): void {
   const statusBarItem = StatusBarItem.getInstance();
+  const inlineMessage = InlineMessage.getInstance();
   runGitBlameCommand()
     .then(async (gitBlameCommandInfo) => {
       if (gitBlameCommandInfo) {
         const commitMessage = gitBlameCommandInfo.gitBlameInfo.summary;
         const jiraIssueKey = getJiraIssueKey(commitMessage);
         statusBarItem.renderStatusBarItem(jiraIssueKey);
-        renderInlineMessage(gitBlameCommandInfo, jiraIssueKey);
+        inlineMessage.renderInlineMessage(gitBlameCommandInfo, jiraIssueKey);
         renderWebview(jiraIssueKey);
       } else {
         statusBarItem.hideStatusBarItem();
-        hideInlineMessage();
+        inlineMessage.hideInlineMessage();
         renderWebview('');
       }
     })
     .catch((error) => {
       console.error('runGitBlameCommand error: ', error.message);
       statusBarItem.hideStatusBarItem();
-      hideInlineMessage();
+      inlineMessage.hideInlineMessage();
       renderWebview('');
     });
 }
